+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to the "watermelon" extension will be documented in this fil
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## [1.8.0]
+- Solved critical bug when clicking "Get code context"
+- Removed daily summary to focus on context
+
 ## [1.7.2]
 
 - Fixes an error preventing the app from opening